move estimateGas.ts to sendShielded.ts
